requirements.txt
langchain
langchain_core
langchain_community
langchain-google-genai
langgraph
langchain_groq
faiss-cpu
streamlit
pymupdf
docx2txt
python-dotenv
tavily-python